python-api/requirements.txt
annotated-types
anyio
blinker
boto3
botocore
certifi
charset-normalizer
click
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama
coloredlogs
fastapi
filelock
Flask
flask-cors
flatbuffers
fsspec
ftfy
git-filter-repo
h11
httptools
huggingface-hub
humanfriendly
idna
imgbeddings
itsdangerous
Jinja2
jmespath
joblib
MarkupSafe
mpmath
networkx
numpy
onnxruntime
opencv-contrib-python
packaging
pillow
protobuf
pydantic
pydantic_core
pyreadline3
python-dateutil
python-dotenv
python-multipart
PyYAML
regex
requests
s3transfer
safetensors
scikit-learn
scipy
setuptools
six
sniffio
starlette
sympy
threadpoolctl
tokenizers==0.13.3
tqdm
transformers
typing_extensions
urllib3
uvicorn
watchfiles
wcwidth
websockets
Werkzeug
exceptiongroup
motor

